p2w-client: Make sure we always update each symbol's state
commit-id:27f276dc

+++ solana/pyth2wormhole/client/src/attestation_cfg.rs
@@ -37,7 +37,7 @@ pub struct AttestationConditions {
 
     /// Trigger attestation if price changes by the specified percentage.
     #[serde(default)]
-    pub price_changed_pct: Option<f32>,
+    pub price_changed_pct: Option<f64>,
 
     /// Trigger attestation if publish_time changes
     #[serde(default)]
